requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pytesseract==0.3.10
paddleocr==2.7.0.3
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
sentence-transformers==2.5.1
numpy==1.26.3
Pillow==10.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pdfplumber==0.11.0 